+++ tailwind.config.js
@@ -7,11 +7,13 @@ module.exports = {
                 primary: {
                     1: "#8000ff",
                     2: "#e000ff",
+                    3: "rgb(30 41 59)",
                 },
                 authform: {
                     1: "#12055c",
                     2: "#42045e",
                 },
+                fadingWhite: "rgba(255,255,255,0.7)",
                 error: "crimson",
                 success: "rgb(20, 220, 60)",
             },